Create january-2021.md
So I put together this report :) Let me know if the tone/info is good. I wanted it to be succinct but not obtuse. I also don't know if I should explicitly state it or if it comes across, but everyone has been really helpful and generous with their time/efforts. In no way is the content of the report any kind of criticism; it's just informational.

I think the easiest way to distribute it would be to share the link to IRC/Slack?