Create january-2021.md (#2)
* Create january-2021.md

So I put together this report :) Let me know if the tone/info is good. I wanted it to be succinct but not obtuse. I also don't know if I should explicitly state it or if it comes across, but everyone has been really helpful and generous with their time/efforts. In no way is the content of the report any kind of criticism; it's just informational.

I think the easiest way to distribute it would be to share the link to IRC/Slack?

> > I think the easiest way to distribute it would be to share the link to IRC/Slack? My other idea was to tag folks here, but I think that might be overkill?
> 
> I agree tagging people on the PR is probably not ideal. Broadcasting the link and letting people consume it if they want seems to make sense. IRC + the mailing list seems like a good plan. Maybe also a tweet?

Agree, and we may as well tweet it :)